Updated pause menus and added the functions
I have not tested this script. Also, you will have to change the name of
this script to get rid of the "(2)."

You're also better off copy and pasting this script and pasting it over
the orginal, otherwise you might have to replace all of the game
objects.

Set the menu up and when it's set up i'll get your version of the game,
and fix it if it's broken.